requirements.txt
easydict==1.9
fishervector==0.1
joblib==1.0.1
matplotlib==3.3.4
numpy==1.20.1
opencv_python==4.5.5.62
pandas==1.2.4
Pillow==9.2.0
PyQt5==5.15.7
scikit_learn==1.1.2
seaborn==0.11.1
torch==1.9.0
torchvision==0.10.0
tqdm==4.59.0
